+++ client/src/app/components/adminCC-dashboard/clinic-listing/clinic-listing.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ClinicService } from 'src/app/services/clinic.service';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
 	selector: 'app-clinic-listing',
@@ -13,8 +14,9 @@ export class ClinicListingComponent implements OnInit {
 	navigationSubscription: any;
 
 	constructor(private clinicService: ClinicService,
-		private router: Router,
-		private activatedRoute: ActivatedRoute) {
+		public router: Router,
+		private activatedRoute: ActivatedRoute,
+		private confirmationDialogService: DialogService) {
 		this.navigationSubscription = this.router.events.subscribe((e: any) => {
 			if (e instanceof NavigationEnd) {
 				this.getClinics();
@@ -23,6 +25,7 @@ export class ClinicListingComponent implements OnInit {
 	}
 
 	ngOnInit() {
+		console.log(this.router.url);
 	}
 
 	getClinics() {
@@ -33,25 +36,30 @@ export class ClinicListingComponent implements OnInit {
 	}
 
 	showClinicInfo(clinic: any) {
-		this.router.navigate(['../showClinicInfo'], { relativeTo: this.activatedRoute, state: {data: clinic}});
+		this.router.navigate(['showClinicInfo'], { relativeTo: this.activatedRoute, state: {data: clinic}});
 	}
 
 	showClinicAdmins(clinic: any) {
-		this.router.navigate(['../clinicAdmins', { id: clinic.id, name: clinic.name }], { relativeTo: this.activatedRoute });
+		this.router.navigate(['clinicAdmins', { id: clinic.id, name: clinic.name }], { relativeTo: this.activatedRoute });
 	}
 
 	showNewClinicForm() {
-		this.router.navigate(['../addClinic'], { relativeTo: this.activatedRoute });
+		this.router.navigate(['addClinic'], { relativeTo: this.activatedRoute });
 	}
 	deleteClinic(clinic){
-		this.clinicService.deleteClinic(clinic).subscribe(
-			(data) => {
-				this.router.navigate(['../clinics'], { relativeTo: this.activatedRoute });
-			},
-			(error)=>{
-				alert(error);
+		this.confirmationDialogService.confirm('Please confirm', 'Are you sure you want to delete clinic: ' + clinic.name + ' ?', false)
+		.then((confirmed) => {
+			if (confirmed.submited) {
+				this.clinicService.deleteClinic(clinic).subscribe(
+					(data) => {
+						this.router.navigate(['../adminCC'], { relativeTo: this.activatedRoute });
+					},
+					(error)=>{
+						alert(error);
+					}
+				);
 			}
-		);
+		});
 	}
 	ngOnDestroy() {
 		if (this.navigationSubscription) {
